requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
frozenlist==1.4.1
idna==3.10
kiwisolver==1.4.7
magic-filter==1.0.12
matplotlib==3.9.2
multidict==6.1.0
numpy==2.1.1
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.11.1
